+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - File transfer strategy for Azure blob storage
+- File transfer strategy for GCP storage
 
 ### Operations